Fixes #1, implemented a quick and dirty solution.
The solution is based on the idea of using the blog_base as schemaless.
This way both http:// and http:// cases would be covered. Rerwote the
queries to either be using LIKE '%SCHEMALESS_BASE_URL' or changed
existing ones to be able to handle it. Please notice the XXX comments in
getCourseLinkingConnectionsData function, it does solve the problem but
is probably not very efficient. Added an issue for implementing a better
solution.